feature: files reorganized
Se reorganizó los archivos a un solo microservicio, respetando estructura general
