--- persona/src/main/java/com/project/demo/endpoints/personaEndPoints.java
@@ -1,10 +1,0 @@
-package com.project.demo.endpoints;
-
-public class personaEndPoints {
-    public static final String PERSONA_ENDPOINT = "/persona";
-    public static final String PERSONA_CREATE = "/create";
-    public static final String PERSONA_UPDATE = "/update";
-    public static final String PERSONA_GET_ALL = "/getAll";
-    public static final String PERSONA_GET_BY_ID = "/getById/{id}";
-    public static final String PERSONA_DELETE = "/delete/{id}";
-}
