--- persona/src/main/resources/application.properties
@@ -1,17 +1,0 @@
-spring.application.name=demo
-server.port = 8081
-
-spring.jpa.hibernate.naming.physical-strategy=org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
-
-# MySQL
-spring.datasource.url=jdbc:mysql://localhost:3312/DBProyecto-2024
-spring.datasource.username=root
-spring.datasource.password=unu2024
-
-# Configuración del logging para ver los logs en la consola
-logging.level.org.springframework=INFO
-logging.level.com.project.demo.controller=DEBUG
-
-# Configuración adicional de logging
-logging.pattern.console=%d{yyyy-MM-dd HH:mm:ss} - %msg%n
-logging.file.name=logs/app.log
