+++ usuario/src/main/java/com/project/demo/adapters/rest/constants/usuarioEndpoint.java
@@ -1,0 +1,10 @@
+package com.project.demo.adapters.rest.constants;
+
+public class usuarioEndpoint {
+    public static final String USUARIO_ENDPOINT = "/usuario";
+    public static final String USUARIO_CREATE = "/create";
+    public static final String USUARIO_UPDATE = "/update";
+    public static final String USUARIO_GET_ALL = "/getAll";
+    public static final String USUARIO_GET_BY_ID = "/getById/{id}";
+    public static final String USUARIO_DELETE = "/delete/{id}";
+}
