--- usuario/src/main/java/com/project/demo/endpoints/usuarioEndpoint.java
@@ -1,10 +1,0 @@
-package com.project.demo.endpoints;
-
-public class usuarioEndpoint {
-    public static final String USUARIO_ENDPOINT = "/usuario";
-    public static final String USUARIO_CREATE = "/create";
-    public static final String USUARIO_UPDATE = "/update";
-    public static final String USUARIO_GET_ALL = "/getAll";
-    public static final String USUARIO_GET_BY_ID = "/getById/{id}";
-    public static final String USUARIO_DELETE = "/delete/{id}";
-}
