+++ usuario/src/main/java/com/project/demo/infrastructure/repositories/IPersonaRepository.java
@@ -1,0 +1,11 @@
+package com.project.demo.infrastructure.repositories;
+
+import org.springframework.data.repository.CrudRepository;
+import org.springframework.stereotype.Repository;
+
+import com.project.demo.domain.personaModel;
+
+@Repository
+public interface IPersonaRepository extends CrudRepository<personaModel, Integer>{
+    
+}
