--- usuario/src/main/java/com/project/demo/services/IUsuarioService.java
@@ -1,13 +1,0 @@
-package com.project.demo.services;
-
-import java.util.List;
-
-import com.project.demo.model.usuarioModel;
-
-public interface IUsuarioService {
-    public usuarioModel add (usuarioModel model);
-    public usuarioModel update (usuarioModel model);
-    public boolean delete (int id);
-    public usuarioModel getById (int id);
-    public List<usuarioModel> getAll ();
-}
