+++ usuario/src/main/resources/application.properties
@@ -10,7 +10,7 @@ spring.datasource.password=unu2024
 
 # Configuración del logging para ver los logs en la consola
 logging.level.org.springframework=INFO
-logging.level.com.project.demo.controller=DEBUG
+logging.level.com.project.demo.adapters.rest=DEBUG
 
 # Configuración adicional de logging
 logging.pattern.console=%d{yyyy-MM-dd HH:mm:ss} - %msg%n
